requirements.txt
anthropic>=0.49.0
openai>=1.0.0
replicate>=0.30.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
mutagen>=1.47.0
playwright==1.44.0
requests==2.31.0
beautifulsoup4==4.12.3
Pillow>=10.0.0

# web
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
celery[redis]>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
